Revert "virtio: use smp_XX barriers on SMP"
Virtio is using memory barriers to control the ordering of
references to the vrings on SMP systems. When the guest is compiled
with SMP support, virtio is only using SMP barriers in order to
avoid incurring the overhead involved with mandatory barriers.

Lately, though, virtio is being increasingly used with inter-processor
communication scenarios too, which involve running two (separate)
instances of operating systems on two (separate) processors, each of
which might either be UP or SMP.

To control the ordering of memory references when the vrings are shared
between two external processors, we must always use mandatory barriers.

A trivial, albeit potentially sub-optimal, solution is to simply revert
commit d57ed95 "virtio: use smp_XX barriers on SMP".

If this is going to have a negative impact on performance of SMP-based
virtualization use cases, a different approach will be taken.

This reverts commit d57ed95da483418e8b0433da693c9168dd0a2df6.

Conflicts:

	drivers/virtio/virtio_ring.c