remoteproc: remove rproc->lock inside rproc_kick
With the rproc->lock in rproc_kick there is a kernel warning about
possible deadlock between that lock and the lock that is hold when
a new message is notified and the warning is completely valid.

The lock is not really necessary it is to serialize calls to
pm_runtime_get_sync when it is in suspend state. However, as the
number of calls to pm_runtime_get_sync match with the calls to
pm_runtime_put_autosuspend there is no problem and it can be removed.

Signed-off-by: Fernando Guzman Lugo <fernando.lugo@ti.com>