omap: rpmsg: resmgr: add omap resource manager driver
This omap resource manage driver implements the functions to request,
release and get information for resources supported by omap remote
processors and then registers the manager with the generic resource
manage driver. The resources that this driver supports are:

-gptimer
-sdma
-auxclk

Each new omap resource manager device is consider as a new manager and
inside the platform data it should export which omap resources are
valid for that specific manager. So, a manager can say it does not want
to support gptimers at all, or just few gptimers. The same applies to
all resources. So, each manager can be customized at will to restrict
resources to different remote processors.

Signed-off-by: Fernando Guzman Lugo <fernando.lugo@ti.com>
Signed-off-by: Miguel Vadillo <vadillo@ti.com>