resmgr: add i2c resource to the resource manager
Add the i2c resource so that a remote processor can request a
i2c resource.

Make use of the i2c driver function when available instead of
pm runtime function. The i2c driver should provide function to
keep i2c running in order a remoteproc can access it and also
provide protection if it is going to be shared.

Signed-off-by: Fernando Guzman Lugo <fernando.lugo@ti.com>
Signed-off-by: Miguel Vadillo <vadillo@ti.com>