omap: resmgr: implement constraint functions
Implement constraint functions for:
-iva
-fdif
-iss

Some constraint functions for omap devices are omap specific or even
machine specific and therefore not all are exported. In order to reach
those functions the omap_rpmsg_resmgr driver would have to be built-in
in the kernel avoiding the possibility of building it as a module and
therefore all its dependencies (RPMSG, RPROC). For that reason  we are
creating small omap resmgr  module in mach-omap2 which creates a
platform device to export the constraint functions as pointers to the
upper driver. That way, omap rpmsg resmgr driver can access the
constraint APIs through the function pointers in the pdata from this
new platform device. Doing this everything can still be built as
modules.

Signed-off-by: Fernando Guzman Lugo <fernando.lugo@ti.com>