omap: resmgr: implement constraint functions
Implement constraint functions for:
-iva
-fdif
-iss

Some constraints for omap resources are omap specific and even mach
specific and they are not exported. In order to reach those functions
from a omap rpmsg resmgr driver that can be built as a module, we are
creating a module in mach-omap2 which creates a platform device to export
some constraint function pointers to the upper driver. That way,
omap rpmsg resmgr driver can access the constraint api through the
function pointers in the pdata from this new platform device.

Signed-off-by: Fernando Guzman Lugo <fernando.lugo@ti.com>