remoteproc: add suspend and resume support
Remoteproc module is registering resume/suspend callbacks into the pm
framework. That way when there is a system suspend the suspend callback
will be called and this will back into a new low driver suspend operation
in case of the low driver has suspend/resume support. The same for
resume case.

This way low level remoteproc drivers can add suspend/resume support
for the devices they control. Also these low level drivers needs to
manage the case when the remoteproc is suspended and the kick callback
is executed. Low level driver needs to provide a safe way to do that.
Most likely the kick should be blocked until resume is done.

Signed-off-by: Fernando Guzman Lugo <fernando.lugo@ti.com>