remoteproc: add suspend and resume support
Remoteproc module now is registering resume/suspend callbacks into
the pm framework using the rproc_class. That way when there is a system
suspend the suspend callback will be called and this callback will check
if the low level driver has a suspend operation that would be called in
order to put the rproc in suspend state (and of course it should save
the  context). The same applies for the resume case.

If the low level driver does not have suspend/resume operations that means
it does not support system syspend, and the low driver is not affect by
the suspend and resume transitions.

In order to avoid possible issues or race conditions sending messages
when the rproc is suspending or suspended, all the kick functions will
be blocked when there is a system suspend running or when the rproc
is suspended. The kick function will be unblocked once the system
resume for that rproc is executed. In the future if some platform does
not support or does not want a blocking kick function, a new field in
the rproc strucure can be added to state if the kick function can be
block or not.

Signed-off-by: Fernando Guzman Lugo <fernando.lugo@ti.com>