remoteproc: add autosuspend support
Going to suspend and resuming a remote processor is a heavy task with a
big latency. For that reason, remoteproc should go to suspend only after
some time of inactivity. Remoteproc module is using autosuspend feature
of PM runtime framework to accomplish that. Once it is autosuspended the
remote processor will be waken up every time the host needs to send it a
message, it is done in the rproc_kick function.

The autosuspend timeout comes from the firmware, in case of the firmware
does not provide an autosuspend timeout that means that remoteproc does
not support autosuspend. Therefore, autosuspen will be disabled in that
case.

In case of autosuspend support, PM runtime framework allows to change
the autosuspend timeout at runtime using a sysfs entry available to each
remote processor.

Example:
echo 4000 > /sys/devices/platform/omap/omap-rproc.1/power/autosuspend_delay_ms

Signed-off-by: Fernando Guzman Lugo <fernando.lugo@ti.com>