omap: remoteproc: handle the vq msg in a kthread
Introduce a kthread to handle the vq messages
arriving to the omap_remoteproc cb, this will let
us handle the messages in a parallel way.

This is also helping with a ugly trace about a possible deadlock when
rproc apis are call inside a rpmsg callback. And as when the rproc
will be added as a resource in the resource manager that trace will
appear due to the rproc lock and the mbox lock are taken in different
order.

Signed-off-by: Miguel Vadillo <vadillo@ti.com>
Signed-off-by: Fernando Guzman Lugo <fernando.lugo@ti.com>