omap: iommu: deny clkdm idle when iommu fault is not handle until shutdown
When and iommu fault happen and it could not be handled that means a
valid physical address could not be found we cannot ack the status bit
of the interrupt. Therefore, we need to deny clkdm idle for the clock of
the iommu until the ack can be done. Otherwise, if there is no
clkdm activity (that can happend if the processor associated with the
processor is turned off) the clkdm will try to go to idle and it will
fail and get stuck in transition state because of the pending interrupt.
In order to avoid that scenario we deny idle in the mmufault ISR and then
in the shutdown of the iommu we ack the interrupt and then allow idle
again.

Signed-off-by: Fernando Guzman Lugo <fernando.lugo@ti.com>