rpmsg: resmgr: omap: request regulator based on ids
There can be any number of different regulators that needs to be
controlled from the remote processor and these are specific to
each board. The corresponding mach layer provides a new op to get
the desired regulator data, and this can looked up using a platform
agnostic id.

The specific rprm regulator data has provision to identify whether
a regulator voltage is fixed or variable, and this info is used
accordingly in calling the regulator_set_voltage API.

Signed-off-by: Fernando Guzman Lugo <fernando.lugo@ti.com>
Signed-off-by: Subramaniam Chanderashekarapuram <subramaniam.ca@ti.com>
Signed-off-by: Suman Anna <s-anna@ti.com>