rpmsg: resmgr: omap: enable iss optional clock
The Camera usecases require the ISS optional clock to be enabled
for complete functionality. This optional clock drives the CSI
interfaces required for the camera sensors.

The ISS request and release functions now properly enable and
disable the optional clock using the clock API directly and using
the name retrieved from the platform data.

Signed-off-by: Fernando Guzman Lugo <fernando.lugo@ti.com>
Signed-off-by: Subramaniam Chanderashekarapuram <subramaniam.ca@ti.com>
Signed-off-by: Suman Anna <s-anna@ti.com>