rpmsg: resmgr: omap: request auxclk based on ids
The current rpmsg resmgr implementation for auxclks is based on the
clock names being supplied from the remote processor. The number of
auxclks and the associated clock names can vary from one OMAP version
to another and are defined using clock data on the Linux-side.

This implemenation is changed to use ids so that the remote processor
code can remain agnostic of the clock names. The number and clock
names are published in the mach layer, and the generic rpmsg resmgr
omap layer queries the names associated with an id by using a
auxclk_lookup ops function defined in the MACH layer.

Signed-off-by: Fernando Guzman Lugo <fernando.lugo@ti.com>
Signed-off-by: Subramaniam Chanderashekarapuram <subramaniam.ca@ti.com>
Signed-off-by: Suman Anna <s-anna@ti.com>