arm: omap: devices: add support for registering rprm regulators
The regulators are usually defined in the board file, and support
has been added to publish the regulator data relevant to the
remote processor resource manager. This data would be used by the
generic rpmsg resmgr driver layer to request/release and control
a regulator on behalf of the remote processor through the rpmsg
resmgr framework.

Signed-off-by: Suman Anna <s-anna@ti.com>
Signed-off-by: Subramaniam Chanderashekarapuram <subramaniam.ca@ti.com>
Signed-off-by: Fernando Guzman Lugo <fernando.lugo@ti.com>

Conflicts:

	arch/arm/mach-omap2/devices.c