remoteproc: add rproc_error_reporter function to notify falta errors
This patch is exporting the rproc_error_reporter function which can be
used to report fatal error to the remoteproc core. This function is
specially thought to be called by low level remoteproc driver code in
case of detecting a fatal error (remoteproc is not functional anymore).
Using this function from another driver should be analyze very carefully,
must of the time that will be considered as an error.

rproc_error_reporter function can be called from any context that is, it
can be called from atomic context without any problem. The reporter
function it creating a new thread (workqueue work) in charge of handling
the error (if possible). This new thread is created the reporter function
is only in charge of reporting the issue and it does not need to wait
until the issue is handle. Also, this allow to call rproc_error_reporter
function from any context. Moreover, this avoid any deadlock condition
which could happen if the rproc_error_reporter function is called from
any function which is indirectly holding any rproc lock.

The reporter function is scheduling the error handler task. This task
is thought to have some features like:

-remoteproc register dump
-remoteproc stack dump
-remoteproc core dump
-Saving of the remoteproc traces in order to be visible after the crash
-Reseting the remoteproc in order to make it functional again (hard recovery)

Right now, it is only printing the fatal error which was detected. The
types of errors are represented by an enum. Right, now I have only added
mmufault and watch dog timer errors. Remoteproc users can add more types
when needed.

Signed-off-by: Fernando Guzman Lugo <fernando.lugo@ti.com>
Signed-off-by: Subramaniam Chanderashekarapuram <subramaniam.ca@ti.com>