remoteproc: add rproc_error_reporter function to notify fatal errors
This patch is exporting the rproc_error_reporter function which can be
used to report fatal errors to the remoteproc core. This function is
specially thought to be called by low-level remoteproc driver code in
case of detecting a fatal error (remoteproc is not functional anymore).
Using this function from another driver should be analyzed very carefully
most of the time that will be considered wrong.

rproc_error_reporter function can be called from any context, that means,
it can be called from atomic context without any problem. The reporter
function is creating a new thread (workqueue work) in charge of handling
the error (if possible).

Creating this new thread is done for two main reasons. First reason is
to be able to call it from atomic context, due to the fact that many
errors trigger an interrupt, so this function can be called directly
from ISR context. Second reason is avoid any deadlock condition which
could happen if the rproc_error_reporter function is called from a
function which is indirectly holding a rproc lock.

The reporter function is scheduling the error handler task. This task
is thought to have some features like:

-remoteproc register dump
-remoteproc stack dump
-remoteproc core dump
-Saving of the remoteproc traces in order to be visible after the crash
-Reseting the remoteproc in order to make it functional again (hard recovery)

Right now, it is only printing the fatal error which was detected. The
types of errors are represented by an enum. I have only added mmufault
and watch dog timer and a default device exception errors. Remoteproc
low-level drivers can add more types when needed.

Signed-off-by: Fernando Guzman Lugo <fernando.lugo@ti.com>
Signed-off-by: Subramaniam Chanderashekarapuram <subramaniam.ca@ti.com>