remoteproc/omap: correct the CMA sizes used for IPU & DSP
The CMA carveout sizes used for OMAP4 targets are stale and
outdated. These do not match the current memory regions
defined in the BIOS-side remote processor platform files.

The IPU carveout sizes are adjusted to the correct sizes
after reducing an additional 1MB region used in older kernel
for secure usecases. The DSP carveout size is adjusted to
actually include the two separate 1MB regions used for the
IPC VRING & IPC DATA sections instead of a single 1MB as used
for both in the previous kernel.

This mismatch didn't cause any major issues because of the
8MB alignment enforced during the definition of the CMA regions.

Signed-off-by: Suman Anna <s-anna@ti.com>