remoteproc/omap: request a timer(s) for remoteproc usage
The usage of an external timer for some purposes is a good
way to lower the power consumption of the remote processor
and allow it to go to a lower state.

Some processors use their internal timers as ticks, this
looks like the right thing to do but this behaviour might
not allow it to go to a lower state since the internal
timer is running.

Provide a way to request a timer(s) to be used by the
remoteproc whenever it is enabled and disable them when
the remoteproc is disabled.

Also for the case of ipu_c0 request gptimer #3 to be used
as tick and set the source ipu_c0 is expecting. For dsp,
gptimer #5 is requested.

Note: if the gptimer is already in use by the time ipu_c0/
dsp is loaded, the processors will fail to boot.

This same mechanism to request a timer(s) can be used for
ipu_c1, watchdog or any other purpose.

Signed-off-by: Miguel Vadillo <vadillo@ti.com>
[ported to 3.6 kernel]
Signed-off-by: Suman Anna <s-anna@ti.com>